requirements.txt
asn1crypto==1.5.1
black==22.6.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dbus-python==1.2.18
decorator==5.1.1
evento==2.0.1
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
http-ece==1.1.0
httplib2==0.20.4
idna==3.3
msgpack==1.0.4
mypy-extensions==0.4.3
omxplayer-wrapper==0.3.3
oscrypto==1.3.0
pathlib2==2.3.7.post1
pathspec==0.9.0
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
git+https://github.com/louisliv/push_receiver.git
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
